MULE-8382: Can't uncompress zip files containing no entries for folders
Conflicts:
	core/src/test/java/org/mule/util/FileUtilsTestCase.java

+++ core/src/main/java/org/mule/util/FileUtils.java
@@ -312,6 +312,12 @@ public class FileUtils extends org.apache.commons.io.FileUtils
                 }
                 else
                 {
+                    File file = new File(directory, entry.getName());
+                    if (!file.getParentFile().exists() && !file.getParentFile().mkdirs())
+                    {
+                        throw new IOException("Unable to create folders for zip entry: " + entry.getName());
+                    }
+
                     InputStream is = zip.getInputStream(entry);
                     OutputStream os = new BufferedOutputStream(new FileOutputStream(f));
                     IOUtils.copy(is, os);
